+++ R/phylo.modularity.r
@@ -25,7 +25,8 @@
 #'   estimated from the empirically-generated sampling distribution (see details in Adams and Collyer 2019).
 #'
 #' @param A A 3D array (p x k x n) containing Procrustes shape variables for all specimens, or a matrix (n x variables)
-#' @param partition.gp A list of which landmarks (or variables) belong in which partition (e.g. A,A,A,B,B,B,C,C,C)
+#' @param partition.gp A list of which landmarks (or variables) belong in which partition: 
+#' (e.g. A, A, A, B, B, B, C, C, C)
 #' @param CI A logical argument indicating whether bootstrapping should be used for estimating confidence intervals
 #' @param phy A phylogenetic tree of {class phylo} - see \code{\link[ape]{read.tree}} in library ape
 #' @param iter Number of iterations for significance testing
@@ -41,8 +42,8 @@
 #' @return Objects of class "CR" from modularity.test return a list of the following:
 #'    \item{CR}{Covariance ratio: The estimate of the observed modular signal.}
 #'    \item{CInterval}{The bootstrapped 95 percent confidence intervals of the CR, if CI = TRUE.}
-#'    \item{CR.boot}{The bootstrapped CR values, if CI = TRUE
-#'    (For more than two partitions, this is the mean CR of pairwise CRs.)}
+#'    \item{CR.boot}{The bootstrapped CR values, if CI = TRUE. 
+#'    For more than two partitions, this is the mean CR of pairwise CRs.}
 #'    \item{P.value}{The empirically calculated P-value from the resampling procedure.}
 #'   \item{Effect.Size}{The multivariate effect size associated with sigma.d.ratio.}
 #'    \item{CR.mat}{For more than two partitions, the pairwise CRs among partitions.}
@@ -63,81 +64,46 @@
 #' Y.gpa<-gpagen(plethspecies$land)    #GPA-alignment
 #' land.gps<-c("A","A","A","A","A","B","B","B","B","B","B") 
 #' 
-#' MT <- phylo.modularity(Y.gpa$coords, partition.gp=land.gps, phy=plethspecies$phy, 
+#' MT <- phylo.modularity(Y.gpa$coords, partition.gp=land.gps, 
+#' phy=plethspecies$phy, 
 #' CI = FALSE, iter=499)
 #' summary(MT) # Test summary
 #' plot(MT) # Histogram of CR sampling distribution 
-phylo.modularity<-function(A,partition.gp,phy, CI=FALSE, iter=999, seed=NULL, print.progress=TRUE){
+phylo.modularity<-function(A, partition.gp, phy, CI = FALSE, 
+                           iter = 999, seed = NULL, print.progress = TRUE){
   if(any(is.na(A))==T){
     stop("Data matrix contains missing values. Estimate these first (see 'estimate.missing').")  }
   if (!inherits(phy, "phylo"))
     stop("phy must be of class 'phylo.'") 
-  if (length(dim(A))==3){ x<-two.d.array(A)
-           p<-dim(A)[1]; k<-dim(A)[2];n<-dim(A)[3]
-           if(length(partition.gp)!=p){stop("Not all landmarks are assigned to a partition.")}
-           }
-  if (length(dim(A))==2){ x<-A; k <-1; p<-ncol(A)
-           if(length(partition.gp)!=ncol(x)){stop("Not all variables are assigned to a partition.")}
-           }
-  gps<-factor(as.numeric(as.factor(partition.gp)))
-  gps.obs <- as.factor(rep(gps,k,each = k, length=p*k))
-  ngps<-nlevels(gps)
-  Nspec<-num.taxa.X<-nrow(x)
-  namesX<-rownames(x)
-  if (is.null(namesX)){
-    stop("No specimen names in data matrix. Please assign specimen names.")  } 
-  if (length(match(phy$tip.label, namesX)) != num.taxa.X && length(phy$tip.label) < num.taxa.X)
-    stop("Tree is missing some taxa present in the data matrix") 
-  if (length(match(phy$tip.label, namesX)) != num.taxa.X && num.taxa.X < length(phy$tip.label)) 
-    stop("Tree contains some taxa not present in present in the data matrix")  
-  phy.parts<-phylo.mat(x,phy)
-  invC<-phy.parts$invC; D.mat<-phy.parts$D.mat
-  if(!is.null(seed) && seed=="random") seed = sample(1:iter, 1)
-  if (length(dim(A))==2){
-    CR.obs<-CR.phylo(x,invC,gps.obs)
-    if(ngps > 2) CR.mat <- CR.obs$CR.mat else CR.mat <- NULL
-    CR.obs <- CR.obs$CR
-    if(print.progress) CR.rand <- apply.phylo.CR(x,invC, gps, k, iter=iter, seed=seed) else
-      CR.rand <- .apply.phylo.CR(x,invC, gps, k, iter=iter, seed=seed) 
-    p.val <- 1-pval(CR.rand)  #b/c smaller values more significant
-    if (p.val==0){p.val<-1/(iter+1)}
-    Z <- effect.size(CR.rand, center=TRUE) 
-    if(CI=="TRUE"){
-#      CR.boot<- boot.CR(x, gps, k,iter=iter, seed=seed)
-      CR.boot<- boot.phylo.CR(x, invC, gps, k,iter=iter, seed=seed)
-      CR.CI<-quantile(CR.boot, c(.025, .975)) 
-    }
-    if(CI=="FALSE"){
-      CR.boot <- NULL
-      CR.CI <- NULL
-    }
-  }
-  if (length(dim(A))==3){
-    angle <- seq(0,89.95,.05)
+  if (length(dim(A))==3){ 
+    p<-dim(A)[1]; k<-dim(A)[2];n<-dim(A)[3]
+    gps<-as.factor(partition.gp)
+    gps.obs <- as.factor(rep(gps,k,each = k, length=p*k))
+    angle <- seq(0,89.95,0.05)
     if(k==2){
       rot.mat<-lapply(1:(length(angle)), function(i) matrix(c(cos(angle[i]*pi/180),
-              sin(angle[i]*pi/180),-sin(angle[i]*pi/180),cos(angle[i]*pi/180)),ncol=2))      
+                                                              sin(angle[i]*pi/180),-sin(angle[i]*pi/180),cos(angle[i]*pi/180)),ncol=2))      
     }
     if(k==3){
       rot.mat<-lapply(1:(length(angle)), function(i) matrix(c(cos(angle[i]*pi/180),
-                sin(angle[i]*pi/180),0,-sin(angle[i]*pi/180),cos(angle[i]*pi/180), 0,0,0,1),ncol=3))      
+                                                              sin(angle[i]*pi/180),0,-sin(angle[i]*pi/180),cos(angle[i]*pi/180), 0,0,0,1),ncol=3))      
     }
     Alist <-lapply(1:n,function(j) A[,,j]) # convert array to list
     if(print.progress){
       cat("\nFinding the optimal rotation for CR\n")
       pb <- txtProgressBar(min = 0, max = length(rot.mat), initial = 0, style=3) 
-      rotatedCRs <-sapply(1:length(rot.mat), function(j) {
+      rotatedCRs <- sapply(1:length(rot.mat), function(j) {
         r <- rot.mat[[j]]
         rotA <- t(mapply(function(a) matrix(t(a%*%r)), Alist))
         setTxtProgressBar(pb,j)
-        CR.phylo(rotA,invC,gps)$CR
+        quick.CR(rotA, gps=gps.obs)
       })
       close(pb)
     } else {
       rotatedCRs <-sapply(1:length(rot.mat), function(j) {
         r <- rot.mat[[j]]
         rotA <- t(mapply(function(a) matrix(t(a%*%r)), Alist))
-        CR.phylo(rotA,invC,gps)$CR
+        quick.CR(rotA, gps=gps.obs)
       })
     }
     avgCR <- mean(rotatedCRs)
@@ -145,32 +111,27 @@ phylo.modularity<-function(A,partition.gp,phy, CI=FALSE, iter=999, seed=NULL, pr
     optAngle <- angle[angCheck==min(angCheck)]; optAngle<-optAngle[1]
     # Optimal rotation 
     if(k==2) optRot <- matrix(c(cos(optAngle*pi/180),
-             sin(optAngle*pi/180),-sin(optAngle*pi/180),cos(optAngle*pi/180)),ncol=2) else
-              optRot <- matrix(c(cos(optAngle*pi/180),
-               sin(optAngle*pi/180),0,-sin(optAngle*pi/180),cos(optAngle*pi/180), 0,0,0,1),ncol=3)
-    x <- t(mapply(function(a) matrix(t(a%*%optRot)), Alist))
-    if(print.progress) {
-      cat("\nPerforming permutations\n")
-      CR.rand <- apply.phylo.CR(x, invC, gps, k, iter=iter, seed=seed)
-    } else CR.rand <- .apply.phylo.CR(x, invC, gps, k, iter=iter, seed=seed)
-    CR.rand[1] <- CR.obs <- avgCR
-#    if(ngps > 2) CR.mat <- CR(x,gps)$CR.mat else CR.mat <- NULL
-    if(ngps > 2) CR.mat <- CR.phylo(x,invC,gps)$CR.mat else CR.mat <- NULL
-    p.val <- pval(1/CR.rand)  #b/c smaller values more significant
-    Z <- effect.size(CR.rand, center=TRUE) 
-    if(CI=="TRUE"){
-#      CR.boot<- boot.CR(x, gps, k,iter=iter, seed=seed)
-      CR.boot<- boot.phylo.CR(x, invC, gps, k,iter=iter, seed=seed)
-      CR.CI<-quantile(CR.boot, c(.025, .975)) 
-    }
-    if(CI=="FALSE"){
-      CR.boot <- NULL
-      CR.CI <- NULL
-    }
+                                sin(optAngle*pi/180),-sin(optAngle*pi/180),cos(optAngle*pi/180)),ncol=2) else
+                                  optRot <- matrix(c(cos(optAngle*pi/180),
+                                                     sin(optAngle*pi/180),0,-sin(optAngle*pi/180),cos(optAngle*pi/180), 0,0,0,1),ncol=3)
+    x <- t(mapply(function(a) matrix(t(a%*%optRot)), Alist)); rownames(x) <- dimnames(A)[[3]]
+    p<-dim(A)[1]; k<-dim(A)[2];n<-dim(A)[3]
+    if(length(partition.gp)!=p){stop("Not all landmarks are assigned to a partition.")}
   }
-  
-  out <- list(CR=CR.obs, CInterval=CR.CI, CR.boot = CR.boot, P.value=p.val, Z = Z,
-              CR.mat = CR.mat, random.CR = CR.rand,
+  if (length(dim(A))==2){ x<-A; k <-1; p<-ncol(A)
+  if(length(partition.gp)!=ncol(x)){stop("Not all variables are assigned to a partition.")}
+  }
+  phy.parts<-phylo.mat(x,phy)
+  invC<-phy.parts$invC; D.mat<-phy.parts$D.mat
+  one<-matrix(1,nrow(x)); I = diag(1,nrow(x),) 
+  Ptrans<-D.mat%*%(I-one%*%crossprod(one,invC)/sum(invC))
+  x <- Ptrans%*%x
+  if (length(dim(A))==3){x <- arrayspecs(x,p=p,k=k)}
+  res <- modularity.test(x,partition.gp,CI=CI, iter=iter, seed=seed, opt.rot = FALSE,
+                         print.progress=print.progress)
+  out <- list(CR=res$CR, CInterval=res$CInterval, CR.boot = res$CR.boot, 
+              P.value=res$P.value, Z = res$Z,
+              CR.mat = res$CR.mat, random.CR = res$random.CR,
               permutations=iter+1, call=match.call())
   class(out) <- "CR"
   out