+++ NEWS.md
@@ -1,3 +1,31 @@
+# CHANGES IN GEOMORPH VERSION 3.3.2 (Minor Release)
+
+### NEW FEATURES
+* Added Box-Cox transformation to `effect.size`
+* Added an argument to flip axes in `plot.gm.prcomp`
+* Added pairwise r, Z, and P-values to `integration.test` and `phylo.integration`
+
+### BUG FIXES
+* Fixed bug in `compare.CR` so object labels are retained
+* Fixed bug with `plot.bilat.symmetry` that missed changed object output in `bilat.symmetry`.
+* Fixed issue with permutations in multi-module configurations for `integration.test` and `phylo.integration`
+* Fixed error in `plot.gm.prcomp` 3D tree plotting with time.plot = TRUE
+* Fixed error in `compare.evol.rates` for univariate data and permutations
+* Fixed plotting parameter arguments for phylo.par in `plot.gm.prcomp`.
+* Made it possible to assign `summary.gm.prcomp` as an object
+* Fixed plotting issue in `warpRefMesh`
+* Fixed single variable issue in `compare.evol.rates` related to R 4.0.2 changes
+* Added factor labels to output object CR.mat in `modularity.test` & `phylo.modularity`
+* Fixed bug in computations of `phylo.modularity`
+* Fixed order consistency for pairwise calculations in `phylo.modularity` 
+* Allows single landmarks as slider LM in gpagen
+* Fixed labels in `plotRefToTarget`, allowing 3d TPS warp grids to show labels and for txt.pos to be passed to 3d plots		
+* Fixed bug in txt.adj argument for `gridPar`
+
+### OTHER CHANGES
+* Updated `readmulti.nts` function turning it into a wrapper of `readland.nts` thus allowing also multiple dta files to be read in and compiled
+
+
 # CHANGES IN GEOMORPH VERSION 3.3.1 (Patch Release)
 
 ### BUG FIXES